readme tweaks and some code cleanup, bump crate version
