+++ api/rust/suborbital/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "suborbital"
-version = "0.5.0"
+version = "0.6.0"
 authors = ["cohix <connorjhicks@gmail.com>"]
 edition = "2018"
 description = "Suborbital Wasm Runnable API"
