+++ docs/wasm.md
@@ -1,18 +1,18 @@
-# Reactr ❤️ Wasm
+# Reactr ❤️ WebAssembly
 
-Reactr has first-class support for Wasm-packaged Runnables. Wasm is an incredibly useful modern portable binary format that allows multiple languages to be compiled into .wasm modules.
+Reactr has first-class support for WebAssembly-packaged Runnables. Wasm is an incredibly useful modern portable binary format that allows multiple languages to be compiled into .wasm modules.
 
-Wasm support in Reactr is powered by [Wasmer](https://github.com/wasmerio/wasmer-go), the hard work they've done to create a powerful Wasm runtime that is extensible has been very much appreciated, and it's been very cool seeing that project grow.
+Wasm support in Reactr is powered by [Wasmer](https://github.com/wasmerio/wasmer-go), the hard work they've done to create a powerful and extensible Wasm runtime has been very much appreciated, and it's been very cool seeing that project grow.
 
-The current supported languages are Rust and Swift, and the Runnable API is available for each. More languages such as AssemblyScript, Go, and C++ are coming soon!
+The current supported languages are Rust (stable) and Swift (experimental). The Runnable API is available for each. More languages such as AssemblyScript, Go, and C++ are coming soon!
 
 To create a Wasm runnable, check out the [subo CLI](https://github.com/suborbital/subo). Once you've generated a `.wasm` file, you can use it with Reactr just like any other Runnable!
 
-Due to the memory limitations of Wasm, Wasm runners accept bytes (rather than arbitrary input) and return bytes. Reactr will handle the conversion of inputs and outputs automatically. Wasm runners cannot currently schedule other jobs, though support for that is coming.
+A multitude of example Runnables can be found in the [testdata directory](../rwasm/testdata).
 
-Here's how to use Wasm Runnables:
+Due to the memory layout of WebAssembly, Wasm runners accept bytes (rather than arbitrary input) and return bytes. Reactr will handle the conversion of inputs and outputs automatically. Wasm runners cannot currently schedule other jobs.
 
-First, install Reactr's Wasm package `rwasm`:
+To get started with Wasm Runnables, install Reactr's WebAssembly package `rwasm`:
 ```bash
 go get github.com/suborbital/reactr/rwasm
 ```
